requirements.txt
# Core dependencies (compatible with Airflow 2.8.1)
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=1.4.28,<2.0
alembic>=1.11.0

# GitHub API
PyGithub>=2.0.0

# Vector database
qdrant-client>=1.6.0
numpy>=1.24.0
pandas>=2.0.0

# Data processing
python-dateutil>=2.8.0
pytz>=2023.0

# Configuration and logging
pyyaml>=6.0.0
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Phase 2: Vector Embeddings & Knowledge Graph
networkx>=3.2.0

# Phase 3: LLM Integration & Advanced AI Features
openai>=1.3.0
anthropic>=0.7.0

# Additional dependencies for data collection
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0
praw>=7.7.0
stackapi>=0.2.0
linkedin-api>=2.0.0
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.0
bcrypt>=4.0.0

# Future phases (commented for now)
# langchain==0.0.350
# matplotlib==3.8.2
# seaborn==0.13.0 